assignment3/requirements.txt
imageio
pillow
pygame
scikit-learn